Fix error types and nested error access
Also fix required pattern for currency */+ regex mistake

Instead of passing the entire errors map to each field,
just pass the relevant FieldError object. This is necessary
because the field object doesn't know its own name, so at the
field level we had to do `errors[name].type`. This kind of access
is bad because name can be any string. By accessing the errors map
outside of the input itself, we can access via dot syntax, and the
fields are checked at compile time. Unfortunately the field names
still have to be strings and have to match the final data structure
of the forms, but at least this value is set up in the same place as
the error field access, i.e.

	```
	name="address.zip"
	error={errors.address.zip}
	```