Fix drawer to open/close on narrow screens
This commit is a bit noisy due to indentation level changes.
All the forms required a down tab.

Form container is now spaced to take up a comfortable amount
of the screen width. This means we can just use global left
align, and remove the need for all the wrapping `<Box>` components
just to left align things.

Add app bar to top of form container

Fixes #224