Refactor forms to better separate concerns
I'm trying to separate the PDF filling from the form assembly.

The functions in irsForms/Main complete all calculations and
produce an array of forms, then pdfFiller/fillPdf takes that
array of forms and pushes the data into the actual PDFDocuments.

This way we can create the complete data model for what will be
inserted into the PDF, and probe it for correctness and for display