Refactor for form-hook upgrade
This also eliminates the need for default values at every input.
Instead, default values are managed by the controller,
and in cases where the values need to be updated by force
in the UI (for example to fill forms when an `edit` button is clicked)
the Controller API is used (`reset`).