Remove arbitrary f1040 generator
This sounded cool at the time, but if fast-check spots an error,
it needs to print out the object that was arbitrarily generated.
The f1040 object has a ton of methods on it, so the output is
full of compiled js, leading to confusion. Better to just
have fast-check generate plain objects only.