Add state reset and error for filing status
If user has single saved and then adds a spouse, we show an
error and remind them to select a new status.

If a user has MFJ selected and deletes a spouse, we show an
error and remind them to select a new status.

Form state in these cases is now reset to empty, and this way
react-hook-form properly shows error message and prevents
submitting data.