Add alert before exiting the page if unsaved form
We only warn the user if they have modified the
form at least once. Since we only care about the
first time it was modified, we only invoke the
onChange callback once, and remove the listener
right after to avoid unnecessary event handling.

We modify some tests so that the components that
use the `Prompt` component have a `Router` wrapper,
and thus avoid having errors thrown.