Add transaction csv importer for uncovered 8949 transactions (#933)
* Add transaction list to portfolio converter

Given a list of transactions, build a portfolio of lots, applying FIFO
rules to each portfolio position.

* Add csv import and column field selection

This adds react-data-table-component to efficiently display the
imported data. After csv file is imported, user can select columns
and map them to required fields of Transaction.

* Add UI to mark initial N rows as ignored

* Add portfolio display, import error handling

If the transaction import produces an invalid
portfolio, the erroneous transaction and the last
valid portfolio are displayed. The user is presented
with a message about how to fix the problem if it is
fixable.

* Add importer from preview into data model

* Fix bug with number of required 8949 forms

* Add random data files for testing

* Fix reset button

* fixing parse import

* Fix incorrect import

# Conflicts:
#	src/data/csvImport.ts

* Add different csv imports for jest vs run

* Fix NaN display in firstnrows box per review

* Fix dark mode styling of data table per review

* Add error for unparsable row, make it correctable

* Fix button spacing after csv import, per review

* Factor out large component to separate file

* Fix bug in ready function

Co-authored-by: Aidan Grimshaw <grimshawaidan@gmail.com>