Fix for excess FICA taxes (Schedule 3 line 11)
FICA taxes are calculated per person which means that any excess
amount that can be claimed should be done for the primary and the
spouse when married filling jointly.

The previous logic did not differentiate the w2s between individuals
so the excess amount was calculated to be too high.