Add a new type ValidatedInformation
Remove all situations in forms where we have to double check that
filingStatus is defined, or that primary person information is defined.

Anything that can be an error should be checked before the 1040 is
attempted to build.

Precheck errors before rendering the Summary page. This fixes a
noticeable rendering lag for the final page, as we were generating
the entire 1040 to calculate any errors if they exist.