Try to make Dependent tests less brittle.
The previous test code would look for specific number of elements
in the page or in the validation errors. This was sensitive to the
number of form elements shown and since we are changing that to add
in the age and blindness status the counts were all wrong.

Instead this replaces with a test to see if the length is greater
than 1.