Change the maximum arbitrary social security witholding.
The previous code used a very large number (10,000,000) as the
maximum possible amount of social security tax withheld, however
this is not possible as in real life the amount should be less
than $8500. Because of this large mismatch the test was being
skipped 100% of the time.

This fix creates a new arbitrary maximum of 10,000 which is much
closer to the range of real values and now allows the test to run