Age and blindness redux (#1098)
* Add in age and blindness fields to the Person interface

Replaces the `birthYear` property of `QualifyingInfo` interface
with the `dateOfBirth`

* Turn `TaxPayer` into a generic type

Use a `Date` for the `dateOfBirth` property but have version that
uses a `string` for persistent storage

* Remove ajv-formats.

Originally I thought it would be useful for Date types but it was not being used.

* Fix up schema generation schema references

Schema are automatically generated to reference the type names from
typescript. So for our concrete types we're providing to the schema
generator, we just have to match the names correctly.

* Fix default type parameter

* Fix lingering type mismatches

* Fix validation tests

* Fix action date/string validation mismatches

* Add in migration for age and blindness.

- Use the redux-persist example to create a migration config that
  enables upgrading the store
- Add in a version number to the persisted object to facilitate
  future migrations
- Not quiet working correctly yet at the FSReducer hard codes in the
  name of the previous migration so that loading from file also
  requires that you refresh the browser

* test out migration pattern based on gist

https://gist.github.com/lafiosca/b7bbb569ae3fe5c1ce110bf71d7ee153

* Remove some code that didn't prove fruitful for doing typed migrations

* Try to make Dependent tests less brittle.

The previous test code would look for specific number of elements
in the page or in the validation errors. This was sensitive to the
number of form elements shown and since we are changing that to add
in the age and blindness status the counts were all wrong.

Instead this replaces with a test to see if the length is greater
than 1.

* Fix unnecessary checks flagged by linter

* Handle any migration types

* Repair merge in renamed / deleted worksheet

* Try to work through test failures in Dependent.test.tsx

Still have two errors.

* Uncomment some code in Dependent.test.tsx to overcome unused variable compiler error

* remove console.debug statement from Dependent.test.tsx

* Fix bug caused by reset of form with {}

* Fix initial undefined selection testing bug

* Work on failing dependent tests, simplify tests

* Fix dob field test

* Fix serialization transforms to handle dates anywhere

* Improve DatePicker to show required error

DatePicker does not have its own properties handling required + blank
UI.

Update tests to include Input is required error for datepicker

* Make the default date of birth be black for the tax payer.

Previous to this the initial date of birth was the current date
which would cause confusion if a user didn't change to their actual
birth date. The initial value was already blank on the Spouse and
Dependents so this only needed to change on the primary tax payer

* Add in standard deduction allowances for age and blindness.

The standard deduction amount increases if you (or your spouse)
are over 65 or are blind. this adds in those additional deduction
amounts as well as updating the calculation for the deduction if
someone else can claim you as a dependent.

* Fix failing test for Schedule A

There was a test that assumed certain amounts for the standard
deduction that no longer hold true with the allowances for age and
blindness.

Refactored that test to be more resilient to these changes.

* Remove unused import from Schedule A test

* Remove unused variable in Schedule A tests causing build to fail

Co-authored-by: Zak Patterson <pattersonzak@gmail.com>