Update code for changes in node modules
Most of these changes were due to changes in react-hook-form (or it's
typescript types file). Some of the other ones are probably related to
Typescript checking more things (such as whether attributes are defined
and whether arguments could be undefined or not). One of the changes is
to work around an issue in the PDF library thinking the passed-in
argument isn't a valid type.