Fix investments after January 31, 2022 not being able to be entered (#2045)
The investments entry form had a limitation that the max date was `new
Date(2021, 12, 31)`. Because Javascript months are 0-indexed, this
actually mapped to January 31, 2022. This also meant that investments
bought or sold after this date couldn't be entered.

Fix this by having the max date be today. I think there's anyways logic
to only use investments in the specified tax year for tax calculations.
Also make sure that the date sold field has a date that isn't older than
the date acquired field.

Signed-off-by: Saikrishna Arcot <saiarcot895@gmail.com>